+++ database/migrations/2019_11_17_164812_create_course_teacher_table.php
@@ -14,11 +14,11 @@ class CreateCourseTeacherTable extends Migration
     public function up()
     {
         Schema::create('course_teacher', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('teacher_id');
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('teacher_id');
             $table->foreign('teacher_id')->references('id')->on('teachers');
 
-            $table->unsignedInteger('course_id');
+            $table->unsignedBigInteger('course_id');
             $table->foreign('course_id')->references('id')->on('courses');
             $table->timestamps();
         });
